feat(fe): policy members changes - splitted code into dedicated components/utils/hooks - implemented "delete member" functionality
