+++ apps/back-office/screens/members/components/PoliciesTable/components/PolicyViewDialog/components/ConfirmDeleteMember/index.ts
@@ -1,0 +1,1 @@
+export { ConfirmDeleteMember } from './ConfirmDeleteMember';
