+++ apps/back-office/screens/members/components/PoliciesTable/components/PolicyViewDialog/components/DateCell/index.ts
@@ -1,0 +1,1 @@
+export { DateCell } from './DateCell';
