+++ apps/back-office/screens/members/components/PoliciesTable/components/PolicyViewDialog/components/MemberNameCell/index.ts
@@ -1,0 +1,1 @@
+export { MemberNameCell } from './MemberNameCell';
