+++ apps/back-office/screens/members/components/PoliciesTable/components/PolicyViewDialog/components/TeamProjectCell/index.ts
@@ -1,0 +1,1 @@
+export { TeamProjectCell } from './TeamProjectCell';
