+++ apps/back-office/screens/members/components/PoliciesTable/components/PolicyViewDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './PolicyViewDialog';
